requirements.txt
discord
discord.py[voice]
spotipy
python-dotenv
PyNaCl
yt-dlp
httpx
aiosqlite